rename: utils폴더 안의 파일명을 케밥케이스로 수정
