+++ src/components/common/Highlight/highlight.tsx
@@ -1,5 +1,5 @@
-import { TEAM_DATA } from '../../../utils/teamData';
-import { TeamName } from '../../../utils/teamNameMap';
+import { TEAM_DATA } from '../../../utils/team-data';
+import { TeamName } from '../../../utils/team-name-map';
 import HighlightCard from './highlight-card';
 
 type HighlightProps = {
